// typings/horaro.org-api-v1-events-schedules.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface HoraroOrgApiV1EventsSchedules {
  data?: {
    id: string;
    name: string;
    slug: string;
    timezone: string;
    start: string;
    start_t: number;
    website: string;
    twitter: string;
    twitch: string;
    description: null;
    setup: string;
    setup_t: number;
    updated: string;
    link: string;
    columns: string[];
    items: {
      length: string;
      length_t: number;
      scheduled: string;
      scheduled_t: number;
      data: string[];
      [k: string]: unknown;
    }[];
    links: {
      rel: string;
      uri: string;
      [k: string]: unknown;
    }[];
    [k: string]: unknown;
  }[];
  [k: string]: unknown;
}